# delta: use discovery.Resource wrapper instead of raw any