--- a/pilot/pkg/xds/debuggen.go
+++ b/pilot/pkg/xds/debuggen.go
@@ -22,6 +22,7 @@ import (
 	"net/url"
 	"strconv"
 
+	discovery "github.com/envoyproxy/go-control-plane/envoy/service/discovery/v3"
 	"github.com/golang/protobuf/ptypes/any"
 
 	"istio.io/istio/pilot/pkg/model"
@@ -85,7 +86,7 @@ func NewDebugGen(s *DiscoveryServer, systemNamespace string) *DebugGen {
 // Generate XDS debug responses according to the incoming debug request
 func (dg *DebugGen) Generate(proxy *model.Proxy, push *model.PushContext, w *model.WatchedResource,
 	updates *model.PushRequest) (model.Resources, error) {
-	res := []*any.Any{}
+	res := model.Resources{}
 	var buffer bytes.Buffer
 	if w.ResourceNames == nil {
 		return res, fmt.Errorf("debug type is required")
@@ -116,9 +117,12 @@ func (dg *DebugGen) Generate(proxy *model.Proxy, push *model.PushContext, w *mod
 		buffer.Write(header)
 	}
 	buffer.Write(response.body.Bytes())
-	res = append(res, &any.Any{
-		TypeUrl: TypeDebug,
-		Value:   buffer.Bytes(),
+	res = append(res, &discovery.Resource{
+		Name: resourceName,
+		Resource: &any.Any{
+			TypeUrl: TypeDebug,
+			Value:   buffer.Bytes(),
+		},
 	})
 	return res, nil
 }