XDS introduced a `discovery.Resource` wrapper a while back. This is required for delta XDS.

When using delta, we need to populate the `name` field. To do this, we marshaled the XDS response to inspect the name, which is hacky and extremely bad for performance.

Instead, all generators now return this wrapper. For Sotw, we simply unwrap it.

This of course has a performance cost, but IMO its worth it, only ~1% more memory allocated:
```
name                                     old time/op        new time/op        delta
RouteGeneration/gateways-6                     43.4ms ± 1%        43.4ms ± 1%    ~     (p=1.000 n=4+4)
RouteGeneration/gateways-shared-6               175ms ± 5%         171ms ± 3%    ~     (p=0.343 n=4+4)
RouteGeneration/empty-6                         829µs ± 4%         847µs ± 4%    ~     (p=0.343 n=4+4)
RouteGeneration/tls-6                           853µs ± 1%         858µs ± 5%    ~     (p=0.343 n=4+4)
RouteGeneration/telemetry-6                     839µs ± 3%         844µs ± 3%    ~     (p=0.686 n=4+4)
RouteGeneration/virtualservice-6               1.64ms ± 3%        1.61ms ± 1%    ~     (p=0.200 n=4+4)
RouteGeneration/authorizationpolicy-6          6.92µs ± 2%        7.00µs ± 2%    ~     (p=0.343 n=4+4)
RouteGeneration/peerauthentication-6           7.00µs ± 2%        6.86µs ± 3%    ~     (p=0.114 n=4+4)
ClusterGeneration/gateways-6                   10.7ms ± 3%        10.8ms ± 1%    ~     (p=0.886 n=4+4)
ClusterGeneration/gateways-shared-6            10.7ms ± 2%        10.8ms ± 4%    ~     (p=0.686 n=4+4)
ClusterGeneration/empty-6                      6.43ms ± 3%        6.47ms ± 6%    ~     (p=0.886 n=4+4)
ClusterGeneration/tls-6                        6.35ms ± 4%        6.42ms ± 6%    ~     (p=0.486 n=4+4)
ClusterGeneration/telemetry-6                  5.04ms ± 3%        5.08ms ± 2%    ~     (p=0.686 n=4+4)
ClusterGeneration/virtualservice-6             1.06ms ± 4%        1.05ms ± 2%    ~     (p=0.343 n=4+4)
ClusterGeneration/authorizationpolicy-6        43.6µs ± 2%        44.3µs ± 1%    ~     (p=0.114 n=4+4)
ClusterGeneration/peerauthentication-6         43.0µs ± 4%        44.3µs ± 1%  +2.99%  (p=0.029 n=4+4)
ListenerGeneration/gateways-6                  22.0ms ± 0%        21.6ms ± 2%    ~     (p=0.114 n=4+4)
ListenerGeneration/gateways-shared-6           44.4µs ± 3%        44.3µs ± 6%    ~     (p=0.886 n=4+4)
ListenerGeneration/empty-6                     1.36ms ± 4%        1.33ms ± 3%    ~     (p=0.486 n=4+4)
ListenerGeneration/tls-6                       1.30ms ± 2%        1.27ms ± 2%    ~     (p=0.114 n=4+4)
ListenerGeneration/telemetry-6                 1.65ms ± 1%        1.63ms ± 0%    ~     (p=0.800 n=4+1)

name                                     old kb/msg         new kb/msg         delta
RouteGeneration/gateways-6                      2.67k ± 0%         2.67k ± 0%    ~     (all equal)
RouteGeneration/gateways-shared-6               12.6k ± 0%         12.6k ± 0%    ~     (all equal)
RouteGeneration/empty-6                          32.2 ± 0%          32.2 ± 0%    ~     (all equal)
RouteGeneration/tls-6                            32.2 ± 0%          32.2 ± 0%    ~     (all equal)
RouteGeneration/telemetry-6                      32.2 ± 0%          32.2 ± 0%    ~     (all equal)
RouteGeneration/virtualservice-6                 79.8 ± 0%          79.8 ± 0%    ~     (all equal)
RouteGeneration/authorizationpolicy-6            0.33 ± 0%          0.33 ± 0%    ~     (all equal)
RouteGeneration/peerauthentication-6             0.33 ± 0%          0.33 ± 0%    ~     (all equal)
ClusterGeneration/gateways-6                      297 ± 0%           297 ± 0%    ~     (all equal)
ClusterGeneration/gateways-shared-6               297 ± 0%           297 ± 0%    ~     (all equal)
ClusterGeneration/empty-6                         287 ± 0%           287 ± 0%    ~     (all equal)
ClusterGeneration/tls-6                           282 ± 0%           282 ± 0%    ~     (all equal)
ClusterGeneration/telemetry-6                     212 ± 0%           212 ± 0%    ~     (all equal)
ClusterGeneration/virtualservice-6               29.9 ± 0%          29.9 ± 0%    ~     (all equal)
ClusterGeneration/authorizationpolicy-6          1.10 ± 0%          1.10 ± 0%    ~     (all equal)
ClusterGeneration/peerauthentication-6           1.10 ± 0%          1.10 ± 0%    ~     (all equal)
ListenerGeneration/gateways-6                   1.68k ± 0%         1.68k ± 0%    ~     (all equal)
ListenerGeneration/gateways-shared-6             3.06 ± 0%          3.06 ± 0%    ~     (all equal)
ListenerGeneration/empty-6                       15.1 ± 0%          15.1 ± 0%    ~     (all equal)
ListenerGeneration/tls-6                         10.9 ± 0%          10.9 ± 0%    ~     (all equal)
ListenerGeneration/telemetry-6                   33.1 ± 0%          33.1 ± 0%    ~     (all equal)

name                                     old resources/msg  new resources/msg  delta
RouteGeneration/gateways-6                      1.00k ± 0%         1.00k ± 0%    ~     (all equal)
RouteGeneration/gateways-shared-6                2.00 ± 0%          2.00 ± 0%    ~     (all equal)
RouteGeneration/empty-6                          2.00 ± 0%          2.00 ± 0%    ~     (all equal)
RouteGeneration/tls-6                            2.00 ± 0%          2.00 ± 0%    ~     (all equal)
RouteGeneration/telemetry-6                      2.00 ± 0%          2.00 ± 0%    ~     (all equal)
RouteGeneration/virtualservice-6                 1.00 ± 0%          1.00 ± 0%    ~     (all equal)
RouteGeneration/authorizationpolicy-6            1.00 ± 0%          1.00 ± 0%    ~     (all equal)
RouteGeneration/peerauthentication-6             1.00 ± 0%          1.00 ± 0%    ~     (all equal)
ClusterGeneration/gateways-6                    1.00k ± 0%         1.00k ± 0%    ~     (all equal)
ClusterGeneration/gateways-shared-6             1.00k ± 0%         1.00k ± 0%    ~     (all equal)
ClusterGeneration/empty-6                         410 ± 0%           410 ± 0%    ~     (all equal)
ClusterGeneration/tls-6                           410 ± 0%           410 ± 0%    ~     (all equal)
ClusterGeneration/telemetry-6                     410 ± 0%           410 ± 0%    ~     (all equal)
ClusterGeneration/virtualservice-6                104 ± 0%           104 ± 0%    ~     (all equal)
ClusterGeneration/authorizationpolicy-6          6.00 ± 0%          6.00 ± 0%    ~     (all equal)
ClusterGeneration/peerauthentication-6           6.00 ± 0%          6.00 ± 0%    ~     (all equal)
ListenerGeneration/gateways-6                    2.00 ± 0%          2.00 ± 0%    ~     (all equal)
ListenerGeneration/gateways-shared-6             2.00 ± 0%          2.00 ± 0%    ~     (all equal)
ListenerGeneration/empty-6                       6.00 ± 0%          6.00 ± 0%    ~     (all equal)
ListenerGeneration/tls-6                         6.00 ± 0%          6.00 ± 0%    ~     (all equal)
ListenerGeneration/telemetry-6                   6.00 ± 0%          6.00 ± 0%    ~     (all equal)

name                                     old alloc/op       new alloc/op       delta
RouteGeneration/gateways-6                     17.2MB ± 0%        17.4MB ± 0%  +0.74%  (p=0.029 n=4+4)
RouteGeneration/gateways-shared-6              68.0MB ± 0%        68.0MB ± 0%    ~     (p=0.486 n=4+4)
RouteGeneration/empty-6                         617kB ± 0%         617kB ± 0%  +0.04%  (p=0.029 n=4+4)
RouteGeneration/tls-6                           617kB ± 0%         617kB ± 0%  +0.04%  (p=0.029 n=4+4)
RouteGeneration/telemetry-6                     619kB ± 0%         619kB ± 0%  +0.04%  (p=0.029 n=4+4)
RouteGeneration/virtualservice-6                870kB ± 0%         870kB ± 0%    ~     (p=0.200 n=4+4)
RouteGeneration/authorizationpolicy-6          4.34kB ± 0%        4.47kB ± 0%  +2.95%  (p=0.029 n=4+4)
RouteGeneration/peerauthentication-6           4.34kB ± 0%        4.47kB ± 0%  +2.95%  (p=0.029 n=4+4)
ClusterGeneration/gateways-6                   4.66MB ± 0%        4.78MB ± 0%  +2.76%  (p=0.029 n=4+4)
ClusterGeneration/gateways-shared-6            4.66MB ± 0%        4.78MB ± 0%  +2.76%  (p=0.029 n=4+4)
ClusterGeneration/empty-6                      2.78MB ± 0%        2.83MB ± 0%  +1.89%  (p=0.029 n=4+4)
ClusterGeneration/tls-6                        2.77MB ± 0%        2.82MB ± 0%  +1.90%  (p=0.029 n=4+4)
ClusterGeneration/telemetry-6                  2.23MB ± 0%        2.28MB ± 0%  +2.35%  (p=0.029 n=4+4)
ClusterGeneration/virtualservice-6              473kB ± 0%         486kB ± 0%  +2.82%  (p=0.029 n=4+4)
ClusterGeneration/authorizationpolicy-6        20.0kB ± 0%        20.8kB ± 0%  +3.84%  (p=0.029 n=4+4)
ClusterGeneration/peerauthentication-6         20.0kB ± 0%        20.8kB ± 0%  +3.84%  (p=0.029 n=4+4)
ListenerGeneration/gateways-6                  10.1MB ± 0%        10.1MB ± 0%  +0.00%  (p=0.029 n=4+4)
ListenerGeneration/gateways-shared-6           21.3kB ± 0%        21.5kB ± 0%  +1.20%  (p=0.029 n=4+4)
ListenerGeneration/empty-6                      825kB ± 0%         826kB ± 0%  +0.09%  (p=0.029 n=4+4)
ListenerGeneration/tls-6                        782kB ± 0%         783kB ± 0%  +0.10%  (p=0.029 n=4+4)
ListenerGeneration/telemetry-6                  928kB ± 0%         929kB ± 0%    ~     (p=0.400 n=4+1)

name                                     old allocs/op      new allocs/op      delta
RouteGeneration/gateways-6                       212k ± 0%          213k ± 0%  +0.47%  (p=0.029 n=4+4)
RouteGeneration/gateways-shared-6                896k ± 0%          896k ± 0%    ~     (p=0.229 n=4+4)
RouteGeneration/empty-6                         6.31k ± 0%         6.31k ± 0%  +0.04%  (p=0.029 n=4+4)
RouteGeneration/tls-6                           6.31k ± 0%         6.31k ± 0%  +0.03%  (p=0.029 n=4+4)
RouteGeneration/telemetry-6                     6.35k ± 0%         6.35k ± 0%  +0.03%  (p=0.029 n=4+4)
RouteGeneration/virtualservice-6                10.7k ± 0%         10.7k ± 0%  +0.01%  (p=0.029 n=4+4)
RouteGeneration/authorizationpolicy-6            48.0 ± 0%          49.0 ± 0%  +2.08%  (p=0.029 n=4+4)
RouteGeneration/peerauthentication-6             48.0 ± 0%          49.0 ± 0%  +2.08%  (p=0.029 n=4+4)
ClusterGeneration/gateways-6                    76.2k ± 0%         77.2k ± 0%  +1.32%  (p=0.029 n=4+4)
ClusterGeneration/gateways-shared-6             76.2k ± 0%         77.2k ± 0%  +1.32%  (p=0.029 n=4+4)
ClusterGeneration/empty-6                       41.7k ± 0%         42.1k ± 0%  +0.98%  (p=0.029 n=4+4)
ClusterGeneration/tls-6                         41.2k ± 0%         41.6k ± 0%  +1.00%  (p=0.029 n=4+4)
ClusterGeneration/telemetry-6                   35.2k ± 0%         35.7k ± 0%  +1.16%  (p=0.029 n=4+4)
ClusterGeneration/virtualservice-6              7.80k ± 0%         7.91k ± 0%  +1.33%  (p=0.029 n=4+4)
ClusterGeneration/authorizationpolicy-6           334 ± 0%           340 ± 0%  +1.80%  (p=0.029 n=4+4)
ClusterGeneration/peerauthentication-6            334 ± 0%           340 ± 0%  +1.80%  (p=0.029 n=4+4)
ListenerGeneration/gateways-6                    130k ± 0%          130k ± 0%    ~     (p=0.200 n=4+4)
ListenerGeneration/gateways-shared-6              306 ± 0%           308 ± 0%  +0.65%  (p=0.029 n=4+4)
ListenerGeneration/empty-6                      14.6k ± 0%         14.6k ± 0%  +0.04%  (p=0.029 n=4+4)
ListenerGeneration/tls-6                        14.2k ± 0%         14.2k ± 0%  +0.04%  (p=0.029 n=4+4)
ListenerGeneration/telemetry-6                  15.8k ± 0%         15.8k ± 0%    ~     (p=0.400 n=4+1)
```

Unfortunately we cannot make this flag protected to allow SotW users  to get that 1% improvement without having 100s of lines of code which I don't think is worth it.