@istio-testing 2021-04-30
@howardjohn: The following test **failed**, say `/retest` to rerun all failed tests:

Test name | Commit | Details | Rerun command
--- | --- | --- | ---
integ-telemetry-k8s-tests_istio | 26a1d526cf3a2f58411ebe071c49320b7f1b075a | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/32564/integ-telemetry-k8s-tests_istio/1388150144863571968) | `/test integ-telemetry-k8s-tests_istio`

<details>

Instructions for interacting with me using PR comments are available [here](https://git.k8s.io/community/contributors/guide/pull-requests.md).  If you have questions or suggestions related to my behavior, please file an issue against the [kubernetes/test-infra](https://github.com/kubernetes/test-infra/issues/new?title=Prow%20issue:) repository. I understand the commands that are listed [here](https://go.k8s.io/bot-commands).
</details>
<!-- test report -->